Moved notification channel creation to NotificationManager.
